Refactor season winner role updater: restructure by concern and clarify handler-service split
